server/requirements.txt
absl-py==0.9.0
astroid==2.4.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
environs==8.0.0
Flask==1.1.2
Flask-Cors==3.0.8
google-api-core==1.19.1
google-auth==1.16.1
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.1
mccabe==0.6.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.2
pylint-flask==0.6
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
toml==0.10.1
tqdm==4.46.1
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.9
uWSGI==2.0.18
Werkzeug==1.0.1
wrapt==1.12.1
